migration

+++ app/user/services/user_services.py
@@ -3,10 +3,11 @@ from app.db.engine import new_session
 from app.reminder.model.reminder_model import ReminderModel
 from app.user.model.user_model import UserModel
 from app.task.model.task_model import TaskModel
+from sqlalchemy import BigInteger
 
 class UserServices:
     @classmethod
-    async def add_user(cls,telegram_id:int,language:str = "ru"):
+    async def add_user(cls,telegram_id:BigInteger,language:str = "ru"):
         async with new_session() as session:
             result = await session.execute(select(UserModel).where(UserModel.telegram_id == id))
             user = result.scalars().first()
